Extrais le gestionnaire d'évènements
